+++ CHANGES.rst
@@ -33,6 +33,7 @@ master (in development)
   - Avoid warnings about missing main functions in `rust`
   - Properly resolve relative filenames in `.. include::` directives in `rst`
   - #348: Use `--unix_mode` option in `javascript-gjslint` to get the file name
+  - Puppet Lint messages now include the name of the corresponding check
 
 0.17 (Feb 1, 2014)
 ------------------